Исправить замечания
