--- src/main/java/ru/practicum/shareit/booking/BookingMapper.java
@@ -1,18 +1,0 @@
-package ru.practicum.shareit.booking;
-
-import org.mapstruct.Mapper;
-import org.mapstruct.MappingConstants;
-import ru.practicum.shareit.booking.dto.BookingModify;
-import ru.practicum.shareit.booking.dto.BookingResponse;
-import ru.practicum.shareit.booking.model.Booking;
-
-import java.util.List;
-
-@Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
-public interface BookingMapper {
-    BookingResponse toBookingResponse(Booking booking);
-
-    List<BookingResponse> toBookingResponseList(List<Booking> bookings);
-
-    Booking toBooking(BookingModify bookingModify);
-}
