--- src/main/java/ru/practicum/shareit/item/CommentMapper.java
@@ -1,21 +1,0 @@
-package ru.practicum.shareit.item;
-
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-import org.mapstruct.MappingConstants;
-import ru.practicum.shareit.item.dto.CommentModify;
-import ru.practicum.shareit.item.dto.CommentResponse;
-import ru.practicum.shareit.item.model.Comment;
-
-import java.util.List;
-
-@Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
-public interface CommentMapper {
-
-    @Mapping(source = "author.name", target = "authorName")
-    CommentResponse toCommentResponse(Comment comment);
-
-    List<CommentResponse> toCommentResponseList(List<Comment> comments);
-
-    Comment toComment(CommentModify commentModify);
-}
