+++ src/main/java/ru/practicum/shareit/item/dto/CommentDTO.java
@@ -1,0 +1,9 @@
+package ru.practicum.shareit.item.dto;
+
+import lombok.Data;
+
+@Data
+public class CommentDTO {
+    private Long id;
+    private String text;
+}
