--- src/main/java/ru/practicum/shareit/item/dto/CommentModify.java
@@ -1,9 +1,0 @@
-package ru.practicum.shareit.item.dto;
-
-import lombok.Data;
-
-@Data
-public class CommentModify {
-    private Long id;
-    private String text;
-}
