+++ src/main/java/ru/practicum/shareit/item/mapper/CommentMapper.java
@@ -1,0 +1,17 @@
+package ru.practicum.shareit.item.mapper;
+
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.MappingConstants;
+import ru.practicum.shareit.item.dto.CommentDTO;
+import ru.practicum.shareit.item.dto.CommentResponse;
+import ru.practicum.shareit.item.model.Comment;
+
+@Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
+public interface CommentMapper {
+
+    @Mapping(source = "author.name", target = "authorName")
+    CommentResponse toCommentResponse(Comment comment);
+
+    Comment toComment(CommentDTO commentDTO);
+}
