--- src/main/java/ru/practicum/shareit/request/ItemRequestMapper.java
@@ -1,18 +1,0 @@
-package ru.practicum.shareit.request;
-
-import org.mapstruct.Mapper;
-import org.mapstruct.MappingConstants;
-import ru.practicum.shareit.request.dto.ItemRequestModify;
-import ru.practicum.shareit.request.dto.ItemRequestResponse;
-import ru.practicum.shareit.request.model.ItemRequest;
-
-import java.util.List;
-
-@Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
-public interface ItemRequestMapper {
-    ItemRequestResponse toItemRequestResponse(ItemRequest itemRequest);
-
-    List<ItemRequestResponse> toItemRequestResponseList(List<ItemRequest> itemRequests);
-
-    ItemRequest toItemRequest(ItemRequestModify itemRequestModify);
-}
