--- src/main/java/ru/practicum/shareit/request/ItemRequestRepository.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.request;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import ru.practicum.shareit.request.model.ItemRequest;
-
-public interface ItemRequestRepository extends JpaRepository<ItemRequest, Long> {
-}
