+++ src/main/java/ru/practicum/shareit/request/repository/ItemRequestRepository.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.request.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import ru.practicum.shareit.request.model.ItemRequest;
+
+public interface ItemRequestRepository extends JpaRepository<ItemRequest, Long> {
+}
