--- src/main/java/ru/practicum/shareit/user/UserMapper.java
@@ -1,13 +1,0 @@
-package ru.practicum.shareit.user;
-
-import org.mapstruct.Mapper;
-import ru.practicum.shareit.user.dto.UserModify;
-import ru.practicum.shareit.user.dto.UserResponse;
-import ru.practicum.shareit.user.model.User;
-
-@Mapper(componentModel = "spring")
-public interface UserMapper {
-    UserResponse toUserResponse(User user);
-
-    User toUser(UserModify userModify);
-}
