--- src/main/java/ru/practicum/shareit/user/UserRepository.java
@@ -1,13 +1,0 @@
-package ru.practicum.shareit.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ru.practicum.shareit.user.model.User;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByEmailEqualsIgnoreCase(String email);
-}
