--- src/main/java/ru/practicum/shareit/user/UserService.java
@@ -1,15 +1,0 @@
-package ru.practicum.shareit.user;
-
-import ru.practicum.shareit.user.dto.UserModify;
-import ru.practicum.shareit.user.dto.UserResponse;
-
-public interface UserService {
-    UserResponse get(Long id);
-
-    UserResponse create(UserModify request);
-
-    UserResponse patch(Long id, UserModify request);
-
-    void delete(Long id);
-
-}
