+++ src/main/java/ru/practicum/shareit/user/controller/UserController.java
@@ -1,0 +1,35 @@
+package ru.practicum.shareit.user.controller;
+
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.*;
+import ru.practicum.shareit.user.service.UserService;
+import ru.practicum.shareit.user.dto.UserDTO;
+import ru.practicum.shareit.user.dto.UserResponse;
+
+@RestController
+@RequestMapping(path = "/users")
+@RequiredArgsConstructor
+public class UserController {
+    private final UserService userService;
+
+    @GetMapping("/{id}")
+    public UserResponse get(@PathVariable Long id) {
+        return userService.get(id);
+    }
+
+    @PostMapping
+    public UserResponse create(@Valid @RequestBody UserDTO request) {
+        return userService.create(request);
+    }
+
+    @PatchMapping("/{id}")
+    public UserResponse patch(@PathVariable Long id, @RequestBody UserDTO request) {
+        return userService.patch(id, request);
+    }
+
+    @DeleteMapping("/{id}")
+    public void delete(@PathVariable Long id) {
+        userService.delete(id);
+    }
+}
