+++ src/main/java/ru/practicum/shareit/user/dto/UserDTO.java
@@ -1,0 +1,16 @@
+package ru.practicum.shareit.user.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotNull;
+import lombok.Data;
+import ru.practicum.shareit.annotation.MinimalId;
+
+@Data
+public class UserDTO {
+    @MinimalId
+    private Long id;
+    private String name;
+    @Email(message = "Некорректная электронная почта")
+    @NotNull(message = "Необходимо указать электронную почту")
+    private String email;
+}
