+++ src/main/java/ru/practicum/shareit/user/mapper/UserMapper.java
@@ -1,0 +1,13 @@
+package ru.practicum.shareit.user.mapper;
+
+import org.mapstruct.Mapper;
+import ru.practicum.shareit.user.dto.UserDTO;
+import ru.practicum.shareit.user.dto.UserResponse;
+import ru.practicum.shareit.user.model.User;
+
+@Mapper(componentModel = "spring")
+public interface UserMapper {
+    UserResponse toUserResponse(User user);
+
+    User toUser(UserDTO userDTO);
+}
