+++ src/main/resources/schema.sql
@@ -1,10 +1,3 @@
-DROP TABLE IF EXISTS "users" CASCADE;
-DROP TABLE IF EXISTS "items" CASCADE;
-DROP TABLE IF EXISTS "bookings" CASCADE;
-DROP TABLE IF EXISTS "item_requests" CASCADE;
-DROP TABLE IF EXISTS "comments" CASCADE;
-
-
 CREATE TABLE IF NOT EXISTS users (
     id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
     name VARCHAR(255) NOT NULL,
